MDL-68677 core: Enable prefetching when cachejs is disabled
With the templaterev issue resolved we should enable the prefetch module
when cachejs is enabled to avoid a different experience when cachejs is
enabled vs. disabled.

Previously there was little point in doing this due to a range of other
bugs (string normalisation, and misuse of templaterev).

With these issues resolved we should enabel the caching.