MDL-68729 Admin: Allow lazy-load of admin_setting_configselect
The admin_setting_configselect admin setting already supported lazy
loading, but this was only available for subclasses. This change
means you can use it directly within the admin setting.

The main advantage is for cases where a setting is not generic but
is only ever going to be required in one place, and it creates
unnecessary clutter to make a new subclass.