MDL-68999 moodlenet: Make the default instance name translatable
Admins can change the value as they want / need. Still, it will be nice
if the default value (the name of the default MoodleNet instance) is
controlled by the language pack.

Beside that, it will allow us to more easily change the value should we
need it in the future (via the en_fix process).