MDL-72228 theme_boost: remove class to account for blocks presence.
Course pages use the the drawers layout file, which included a class
to identify whether the page included blocks (`has-blocks`), applying
a width offset to the main region settings menu.

Since 2305b3eb all course-level blocks are hidden away by default in
a drawer, so the `has-blocks` class had the effect of shifting the
action menu away from the intended right hand side of the page.