MDL-72790 block_timeline: Fix button layout and row wrapping on mobile
This ensures all row content wraps nicely on smaller screen sizes, and
forces all action buttons to be displayed in a consistent position
indepdendent of the length of the activity name etc. Buttons are also
now aligned under the text on small screens, to maximise space for the
text content.

Co-authored-by: Michael Hawkins <michaelh@moodle.com>