MDL-72799 behat: Increase allowed size for privacy exports
Because of more moodle.css contents, the total size of the privacy
exports has grown over the allowed size. Just raise it a little bit
more.

Note this commit isn't a problem caused exclusively by this issue,
but as far as this also contributed to the size increase, assigning
the fix to it.