MDL-71696 question: Align new tables with install/upgrade
Amended because, in all cases, install.xml wins and we
have make upgrade.php to match it...

But in the question_versions->status case, that needs to
be NOT_NULL, so upgrade was correct and install.xml has
been changed for it.

Note this has been changed without using the editor because
it has been detected that there are other changes in the xml
file that have been added manually, we'll create a "reconcile"
issue to fix that (mainly whitespace).

So, yes, please, always use the editor.