MDL-66453 uploadusers: Fix behat with classic
This patch fixes the behat failures with classic for the scenario
"Upload users enrolling them on courses and assign category roles".

There is a "Participants" link hidden in the "Site pages" which is
causing these failures.
Apart from this, there was another issue because it was not possible
to navigation with classic (I've added one extra step to visit homepage).