MDL-75776 composer: Move to upstream mink-extension and updates
We were using our fork of the mink extension:

https://github.com/moodlehq/MinkExtension.git

Because they bumped requirements to PHP 7.4 too early for us. Now,
with Moodle 4.1, finally our requirements are also PHP 7.4, so we
can stop using the fork and switch to the upstream one:

https://github.com/FriendsOfBehat/MinkExtension

Note this only can be done in 4.1dev, because this is the first
version using PHP >= 7.4. Older branches will need to continue
using the fork. Also, note that we may need to go back to the
fork if there is any future trouble with the upstream library
not matching our supported PHP versions.

Apart from that, it's a good moment to bump other components to
current ones.

Of course, this has been generated following the instructions:

https://docs.moodle.org/dev/Composer

And using PHP 7.4 (always the min version supported) to generate
the lock file.